WIP: riscv: mark kernel page table entries as global
The G bit in a RISC-V PTE designates a "global" mapping; i.e. a physical
to virtual address mapping that is valid across all contexts. Applying
this bit to kernel mappings offers a small performance benefit, as it
allows these translations to be preserved during a normal TLB flush. For
example, during a context switch, only the mappings related to the
previous process (non-global mappings) will be flushed from the TLB.

This bit isn't applied explicitly to each PTE, rather it is applied to
all level 1 PTEs, as the global bit is defined to apply implicitly to
lower level PTEs.

It was suggested that the feature be made toggle-able, in the interest
of avoiding possible security exploits. Currently, this is tricky due to
the fact that on RISC-V the low-level page table setup is performed in
locore.S. However, a tunable can easily be added in the future after
this has been reworked.

TODO: go re-verify where and how the PTE_G bit is set