gdb: report specific stop reason for watchpoints
The remote protocol allows for implementations to report more specific
reasons for the break in execution back to the client.

As doing this is entirely optional, it is only implemented on amd64 and
i386. It is not as simple to obtain the address that triggered the
watchpoint on other architectures, so simply avoid doing so for now.