Split out dumper allocation from list insertion
The new function dumper_create() can be used to allocate a dumper
without including it in the list of dumpers available to the debugger.
This is desirable for performing live dumps of the system, which
explicitly does not operate in the debugger context.

free_single_dumper() is made public and renamed to dumper_destroy().