riscv: rework CPU identification [1/6]
Make better use of the RISC-V identification CSRs: mvendorid, marchid,
and mimpid.

This code was written before these registers were well-specified, or
even available to the kernel. It currently fails to recognize any CPU or
platform.

Per the privilege specification, mvendorid contains the JEDEC vendor ID,
or zero.

The marchid register denotes the CPU microarchitecture. This is either
one of the globally allocated open-source implementation IDs, or the
field has a custom encoding. Therefore, for known vendors (SiFive) we
can also maintain a list of known marchid values. If we can not give a
name to the CPU but marchid is non-zero, then just print its value in
the report.

The mimpid (implementation ID) could be used in the future to more
uniquely identify the micro-architecture, but it really remains to be
seen how it gets used. For now we just print its value.