riscv: handle newer riscv,isa property strings
Newer versions of QEMU (7.1.0) will append to the riscv,isa property
indicating the presence of pseudo-extensions such as Zfencei. This
breaks some assumptions about how to parse the string, and triggers a
KASSERT about the expected length.

For example, what is generated with a default virt instance:
    cpu@0 {
        ...
        riscv,isa = "rv64imafdch_zicsr_zifencei_zba_zbb_zbc_zbs";
    }

Allocate the buffer for the property dynamically, and stop when we
encounter an underscore. This is a stop-gap to quickly fix booting with
QEMU, and in our CI. Future changes will be made to handle this better
and parse more extensions.