riscv: handle misaligned address exceptions
If this exception is coming from userspace, send the appropriate SIGBUS
to the process. If it's coming from the kernel this is still fatal, but
we can give a better panic message.

With this we handle all defined exception types.

PR:		266109
Found by:	syzkaller
Reported by:	P1umer <p1umer1337@gmail.com>