wakeup: assert the scheduler is running
We want to avoid calling these functions from a panic/debugger context.
The resulting calls to sched_add() can trigger assertions about
recursing on the thread lock, depending on the system state when it
panicked.

These assertions should increase the visibility of shutdown handlers
that do not respect SCHEDULER_STOPPED().