riscv: rework CPU identification [2/6]
Call identify_cpu() earlier for the BSP.

It is advantageous to have knowledge of ISA features as early as
possible. For example, the presence of newer virtual memory extensions
may be useful to pmap_bootstrap().

To achieve this, split out the printf() parts of identify_cpu() into a
separate function, printcpuinfo(). This latter function will be called
later in boot after the console has been initialized.

Plus, a small trick in init_secondary() allows us to avoid interlaced
output from concurrent calls to printcpuinfo() as APs are started.