hwpmc: implement basic riscv support
Implement the MD functions, and a new hwpmc class. Since RISC-V does not
define a complete set of performance monitoring events and hardware that
are available to supervisor-mode software like the kernel, this is not a
"native" class but one built on top of the PMU Extension for the
Supervisor Binary Interface (SBI) [1].

Sampling mode PMCs are unsupported.

[1] https://github.com/riscv-non-isa/riscv-sbi-doc/blob/master/riscv-sbi.adoc#performance-monitoring-unit-extension-eid-0x504d55-pmu