pmcstat: fix duplicate event allocation on CPU 0
Commit b6e28991bf3a modified the allocation path for system scope PMCs
so that the event was allocated early for CPU 0. The reason for this is
so that the PMC's capabilities could be checked, to determine if pmcstat
should allocate the event on every CPU, or just on one CPU in each NUMA
domain. In the current scheme, there is no way to determine this
information without performing the allocation.

This broke the established use-case of log analysis, and so
0aa150775179a was committed to fix the assertion. The result was what
appeared to be functional, but in normal counter measurement pmcstat was
silently allocating two counters for CPU 0.

The problem is that, depending on the particular hardware/event, we
might not be able to allocate the same event twice on a single CPU.

The simplest solution for now is to release the early-allocated PMC once
we have obtained its capabilities.

This partially reverts commit b6e28991bf3aadb74c54c7e6dfc2992b29abae5a.