WIP: vm_page: add man page references to comments
For all functions which have them. This makes it more (but not
perfectly) obvious which functions have associated documentation.

This improves the discoverability of these pages, which of course is a
benefit to readers, but it also helps authors and reviewers recognize
when a change breaks the documented interface.

The competing comment styles are unchanged.