riscv: T-HEAD PBMT support
T-HEAD CPUs provide a spec-violating implementation of page-based memory
types, using PTE bits [63:59]. Aside from the philosophical problems
with this approach, the reality of increasing availability of T-HEAD
platforms which implement this means that we need to consider adapting
to this.

Add support for this "errata". Note that this change is not enough on
its own, but some workaround will be needed for the bootstrap (locore)
page tables as well.

Reviewed by:	jhb
MFC after:	2 weeks
Sponsored by:	The FreeBSD Foundation
Differential Revision:	https://reviews.freebsd.org/D45472