riscv: enable allwinner RTC
(Second attempt.)

For the Allwinner D1 (Nehza) SBC.

This RTC driver is also a clock provider, which registers two fixed
clocks. In all the devices we currently support, the names of the clocks
are present in the "clock-output-names" property of the device tree.
This is not the case for the D1 DTS, as this property does not appear in
upstream. Therefore the clock definitions are statically assigned a
name, which is overridden when specified.

The driver also supports the older A10/A20 models, which don't export
any clocks. Ensure this case is handled properly.