xhci debug: Hacks to get to a working system
With these hacks (some are probably not needed), I am able to use xhci
debug as a serial connection from both loader and the kernel.

The target is a morefine m6 Intel N200 based system with an Alderlake-N
based xhci controller and the host is a 13th Gen Framework 13 (that
part shouldn't matter). The target is a virtual machine on the m6
running in bhyve with pcie passthrough of the xhci controller.

I am using a Datapro xhci debug cable (USB 3.0 Super-Speed A/A
Debugging Cable).

	https://www.datapro.net/products/usb-3-0-super-speed-a-a-debugging-cable.html